Added spawn delay for cops
- Speeding cars get added to a queue
- Cops will spawn based on queue

- Need to work on multiple cars speeding at the same time

+++ VehicleSim2023v021/Pedestrian.java
@@ -29,7 +29,7 @@ public abstract class Pedestrian extends SuperSmoothMover
         if (awake){
             // Check in the direction I'm moving vertically for a Vehicle -- and only move if there is no Vehicle in front of me.
             if (getOneObjectAtOffset(0, (int)(direction * getImage().getHeight()/2 + (int)(direction * speed)), Vehicle.class) == null){
-                setLocation (getX(), getY() + (int)(speed*direction));
+                setLocation (getX(), getY() + (speed*direction));
             }
             if (direction == -1 && getY() < 100){
                 getWorld().removeObject(this);
